Adjusted the data types
As discussed before, these are changes I already made and uploaded as a separate document. I am now just creating it as a pull request to make it easier to track and not have to keep track of the file names etc.

This builds on what Chris did and just corrects the data types.