Adding linear regression and random forest models
I added a linear regression model, then tested it with a confusion matrix as you guys had done, then I tested it against the data to see which players it gave the highest probability of being an all-star. I did a similar thing with the random forest model.

The linear regression may need further narrowing - I limited the variables to those selected by the RFE, but many of them still had a pretty high p-value so we could drop some of them if we wanted.